thread: refactor MPIR layer thread utils
