+++ src/include/mpir_thread.h
@@ -88,23 +88,27 @@ extern MPID_Thread_tls_t MPIR_Per_thread_key;
 #endif /* MPICH_IS_THREADED */
 
 void MPIR_Add_mutex(MPID_Thread_mutex_t * p_mutex);
+void MPIR_Thread_CS_Init(void);
+void MPIR_Thread_CS_Finalize(void);
 
-#if MPICH_THREAD_GRANULARITY == MPICH_THREAD_GRANULARITY__GLOBAL || \
-    MPICH_THREAD_GRANULARITY == MPICH_THREAD_GRANULARITY__POBJ   || \
-    MPICH_THREAD_GRANULARITY == MPICH_THREAD_GRANULARITY__VCI
-extern MPID_Thread_mutex_t MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX;
-#endif
+#if defined(MPICH_IS_THREADED)
+
+MPIR_EXTERN MPID_Thread_mutex_t MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX;
 
 #if MPICH_THREAD_GRANULARITY == MPICH_THREAD_GRANULARITY__POBJ
-extern MPID_Thread_mutex_t MPIR_THREAD_POBJ_HANDLE_MUTEX;
-extern MPID_Thread_mutex_t MPIR_THREAD_POBJ_MSGQ_MUTEX;
-extern MPID_Thread_mutex_t MPIR_THREAD_POBJ_COMPLETION_MUTEX;
-extern MPID_Thread_mutex_t MPIR_THREAD_POBJ_CTX_MUTEX;
-extern MPID_Thread_mutex_t MPIR_THREAD_POBJ_PMI_MUTEX;
+MPIR_EXTERN MPID_Thread_mutex_t MPIR_THREAD_POBJ_HANDLE_MUTEX;
+MPIR_EXTERN MPID_Thread_mutex_t MPIR_THREAD_POBJ_MSGQ_MUTEX;
+MPIR_EXTERN MPID_Thread_mutex_t MPIR_THREAD_POBJ_COMPLETION_MUTEX;
+MPIR_EXTERN MPID_Thread_mutex_t MPIR_THREAD_POBJ_CTX_MUTEX;
+MPIR_EXTERN MPID_Thread_mutex_t MPIR_THREAD_POBJ_PMI_MUTEX;
 
 #define MPIR_THREAD_POBJ_COMM_MUTEX(_comm_ptr) _comm_ptr->mutex
 #define MPIR_THREAD_POBJ_WIN_MUTEX(_win_ptr)   _win_ptr->mutex
-#endif
+
+#elif MPICH_THREAD_GRANULARITY == MPICH_THREAD_GRANULARITY__VCI
+MPIR_EXTERN MPID_Thread_mutex_t MPIR_THREAD_POBJ_HANDLE_MUTEX;
+
+#endif /* MPICH_THREAD_GRANULARITY */
 
 #ifdef MPICH_THREAD_USE_MDTA
 
@@ -167,8 +171,6 @@ MPL_STATIC_INLINE_PREFIX void MPIR_Thread_sync_wait(MPIR_Thread_sync_t * sync)
 
 #endif /* MPICH_THREAD_USE_MDTA */
 
-#if MPICH_THREAD_GRANULARITY == MPICH_THREAD_GRANULARITY__VCI
-extern MPID_Thread_mutex_t MPIR_THREAD_POBJ_HANDLE_MUTEX;
-#endif
+#endif /* MPICH_IS_THREADED */
 
 #endif /* MPIR_THREAD_H_INCLUDED */
