+++ src/mpi/init/mpi_init.h
@@ -49,8 +49,6 @@ void MPII_init_thread_failed_exit_cs(void);
 void MPII_finalize_thread_and_enter_cs(void);
 void MPII_finalize_thread_and_exit_cs(void);
 void MPII_finalize_thread_failed_exit_cs(void);
-void MPIR_Thread_CS_Init(void);
-void MPIR_Thread_CS_Finalize(void);
 
 int MPII_init_global(int *p_thread_required);
 int MPII_post_init_global(int thread_provided);
