+++ src/util/Makefile.mk
@@ -11,6 +11,7 @@ mpi_core_sources +=   \
     src/util/mpir_assert.c     \
     src/util/mpir_cvars.c      \
     src/util/mpir_pmi.c        \
+    src/util/mpir_thread.c     \
     src/util/mpir_handlemem.c  \
     src/util/mpir_strerror.c   \
     src/util/mpir_localproc.c
